test: setup `venv` once per test suite
Use parent `conftest.py` to setup a venv from which all tests will be run
