+++ upgrade/tests/upgrade_package/conftest.py
@@ -1,40 +1,20 @@
 import os
-import shutil
-import sys
+
 import pytest
 from mock import patch
 from pathlib import Path
-from upgrade.scripts.utils import is_windows
+
 from upgrade.scripts.upgrade_python_package import pip, run
+from ..conftest import VENV_PATH
 
 CLOUDSMITH_URL = os.environ.get("CLOUDSMITH_URL", False)
 
 REPOSITORY_WHEELS_PATH = Path(__file__).parent.parent / "repository"
-VENV_PATH = Path(__file__).parent / "venv"
-original_executable = sys.executable
-
-
-def pytest_configure(config):
-    if is_windows():
-        python_path = Path("Scripts", "python.exe")
-    else:
-        python_path = Path("bin", "python")
-    sys.executable = str(VENV_PATH / python_path)
-    sys.modules["oll"] = None
 
 
 @pytest.fixture
 def wheels_dir():
     return REPOSITORY_WHEELS_PATH
-
-
-@pytest.fixture(scope="session", autouse=True)
-def setup_venv():
-    if VENV_PATH.exists():
-        shutil.rmtree(VENV_PATH)
-    os.makedirs(VENV_PATH, exist_ok=True)
-    run(original_executable, "-m", "venv", str(VENV_PATH))
-    yield
 
 
 @pytest.fixture(autouse=True)
